Update PmergeMe class: change getData to return size_t and add out-of-range check

+++ cpp09/ex02/PmergeMe.hpp
@@ -6,7 +6,7 @@
 /*   By: pnaessen <pnaessen@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/22 17:43:48 by pnaessen          #+#    #+#             */
-/*   Updated: 2025/10/23 12:14:50 by pnaessen         ###   ########lyon.fr   */
+/*   Updated: 2025/10/23 12:58:18 by pnaessen         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ class PmergeMe
 
 		PmergeMe &	operator=(PmergeMe const& assign);
 		void processParsing(int argc, char **argv);
-		int getData(int idx);
+		size_t getData(size_t idx);
 
 };
